requirements.txt
python-decouple==3.6
pandas
-e file:./#egg=non_sequential_split